requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
openpyxl==3.1.2
Pillow==10.1.0
tqdm==4.66.2
scikit-learn==1.3.2

# Segmentation dependencies
numpy>=1.21.0
pydicom>=2.3.0
matplotlib>=3.5.0
scipy>=1.7.0
scikit-image>=0.20.0
SimpleITK>=2.2.0
pylibjpeg
pylibjpeg-libjpeg
pylibjpeg-openjpeg
gdcm

# GPU acceleration
cupy-cuda12x>=12.0.0  # For CUDA 12.x
# Alternative for different CUDA versions:
# cupy-cuda11x>=11.0.0  # For CUDA 11.x
# cupy-cuda10x>=10.0.0  # For CUDA 10.x